+++ tomcat/src/main/java/com/techcourse/controller/Controller.java
@@ -1,4 +1,4 @@
-package org.apache.coyote;
+package com.techcourse.controller;
 
 import org.apache.coyote.http11.HttpRequest;
 import org.apache.coyote.http11.HttpResponse;
