+++ tomcat/src/main/java/com/techcourse/session/Session.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package com.techcourse.session;
 
 import java.util.HashMap;
 import java.util.Map;
