+++ tomcat/src/main/java/org/apache/coyote/util/FileReader.java
@@ -1,5 +1,6 @@
 package org.apache.coyote.util;
 
+import com.techcourse.exception.ResourceNotFoundException;
 import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
@@ -9,7 +10,6 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.List;
 import java.util.Objects;
-import org.apache.coyote.exception.ResourceNotFoundException;
 
 public class FileReader {
